Cap nhat cac chuc nang cua Order Detail

+++ src/main/java/com/project/WebTapGym/models/OrderDetail.java
@@ -24,11 +24,11 @@ public class OrderDetail {
     private Product product;
 
     @Column(name = "price", nullable = false)
-    private int price;
+    private Float price;
 
     @Column(name = "number_of_products", nullable = false)
     private int numberOfProducts;
 
     @Column(name = "total_money", nullable = false)
-    private int totalMoney;
+    private Float totalMoney;
 }
